[util/encodeFloatToFixPoint] extract toFixed
