+++ packages/util/src/toFixPoint.ts
@@ -1,7 +1,7 @@
 import assert from 'assert';
 import BN from 'bn.js';
 
-import { assertLength } from './common';
+import {assertLength, fractionalToRadix2, safeIntegerToRadix2} from './common';
 
 export interface ToFixPointFn {
   (num: number): BN;
@@ -63,25 +63,4 @@ const toFixed = function(integers: string, fractions: string, integer_count: num
    const bits = integers.padStart(integer_count, '0').padEnd(integer_count + fractions_count, '0');
    return new BN(bits, 2);
   }
-}
-
-
-/**
- * Our fixed point integer values go until I64, which means that it may be > 53 bits.
- * So we can't just parse the whole number with `parseInt`, as an overflow would occur.
- *
- * @param num Integer number with base 10 radix
- */
-export const safeIntegerToRadix2 = function(num: string): string {
-  return new BN(num, 10).toString(2)
-}
-
-/**
- * Transforms the fractional value of a number to base 2.
- *
- * @param num Integer number with base 10 radix
- */
-export const fractionalToRadix2 = function(num: string): string {
-  return parseFloat('0.' + num).toString(2)
-      .split('.')[1];
 }
